// src/modules/core/user/translate.enum.ts
export enum UserKeys {
    USER_NOT_FOUND = "user.USER_NOT_FOUND",
    USER_PASSWORD_NOT_MATCH = "user.USER_PASSWORD_NOT_MATCH",
    UPDATED_USER_PROFILE_SUCCESSFULLY = "user.UPDATED_USER_PROFILE_SUCCESSFULLY",
    EMAIL_VERIFICATION_CODE_SENT_SUCCESSFULLY = "user.EMAIL_VERIFICATION_CODE_SENT_SUCCESSFULLY",
    VERIFICATION_CODE_NOT_VALID = "user.VERIFICATION_CODE_NOT_VALID",
    VERIFICATION_CODE_EXPIRED = "user.VERIFICATION_CODE_EXPIRED",
    UPDATED_USER_EMAIL_SUCCESSFULLY = "user.UPDATED_USER_EMAIL_SUCCESSFULLY",
    EMAIL_EXISTS = "user.EMAIL_EXISTS",
    EMAIL_INVITED_AS_INVESTOR = "user.EMAIL_INVITED_AS_INVESTOR",
    DELETION_ERROR = "user.DELETION_ERROR",
}